+++ src/index.ts
@@ -264,7 +264,7 @@ export interface PurchasesTransaction {
   /**
    * RevenueCat Id associated to the transaction.
    */
-  revenuecatId: string;
+  revenueCatId: string;
   /**
    * Product Id associated with the transaction.
    */